requirements.txt
Adafruit-Blinka==8.54.0
adafruit-circuitpython-apds9960==3.1.14
adafruit-circuitpython-busdevice==5.2.11
adafruit-circuitpython-connectionmanager==3.1.3
adafruit-circuitpython-register==1.10.2
adafruit-circuitpython-requests==4.1.9
adafruit-circuitpython-sht4x==1.0.22
adafruit-circuitpython-typing==1.11.2
Adafruit-PlatformDetect==3.77.0
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
certifi==2025.1.31
charset-normalizer==3.4.1
# Editable install with no version control (fireplace==0.1)
-e /home/hotone/Documents/Fireplace/FireplaceThermostat
idna==3.10
numpy==2.2.3
pandas==2.2.3
pyftdi==0.56.0
pygame==2.6.1
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2025.1
pyusb==1.3.1
requests==2.32.3
rpi-backlight==2.6.0
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
schedule==1.2.2
six==1.17.0
sysv-ipc==1.1.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
